Update README.md
However, there's a beacon of hope in the form of Fresh Heads Lice Removal. This specialized service is where lice meet their match. Fresh Heads employs a team of trained professionals equipped with cutting-edge techniques and natural products to safely and effectively eliminate lice and nits. Their commitment to providing a lice-free environment extends to offering education and prevention tips to families. With a focus on non-toxic methods and a dedication to eradicating lice, Fresh Heads Lice Removal is the go-to destination for those dealing with these pesky parasites. Say goodbye to lice-related stress and hello to a lice-free life with Fresh Heads.
https://beautysalonorbit.com/fresh-heads-lice-removal/